feat: add incomplete AmigaOS support
Unfortunately, tests based on FS-UAE did not succeed due to the
serial device not emitting a signal on receiving data from the
calculator.

This commit is made so that testing may be done on other platforms,
and an initial version of the AmigaOS support is at least present
in the develop branch.