fix: add an option to build and install CaS
In order to avoid building a program we won't be installing,
this commit adds an option to toggle or not CaS program
build and installation, instead of the previously commented
installation only.

This is useful for packaging systems that check the existence
of built but uninstalled programs, such as Homebrew.