feat: add SEND_FILE_FLAG_DELETE flag
This flag, if used along with SEND_FILE_FLAG_OPTIMIZE,
allows one to reclaim space occupied by the previous
version of the file for the new one. It allows sending
a 1 MB file with the same name multiple times in a row,
which is more practical for add-in development.

This is a separate flag from FORCE since it causes the file,
if present, to be explicitely deleted rather than relying on
the p7 server's behaviour, which is not guaranteed by the
protocol.

Closes: #29