fix: use removal relations to find volume
In ScreenRecv mode, as opposed to Projector mode, Windows
does not present the volume to be used as a bus relations
device to the disk drive related to the USB device.

However, in both cases, the volume is present as a device
in the removal relations at least on Win11; therefore, we
use those instead.

Closes: #18