fix: re-use asynchronous read operations on Windows
This commit makes it so that we re-use the previous asynchronous
read on Windows serial and CESG502 links, in order to avoid crashing
while the link is active on Windows XP when calling CancelIo().

It is not a complete fix since calling CancelIo() when closing, or
not calling CancelIo() then re-running the program, still causes a
crash when using CESG502. The driver is probably to blame, but there
may still be options on our side to explore.