fix: increase p7 init response timeout to 800ms
Some older models such as the fx-9860G Slim seem not to be
able to answer in time sometimes within 500ms, causing a shift
in acknowledgement types which corrupt the link until the cable
is unplugged then plugged.

This commit increases the response timeout for Protocol 7.00
initial check packets from 500ms to 800ms in all situations
(automatic detection and explicit Protocol 7.00 initialization)
to make this scenario more unlikely.

Related to: #44